Only check on if admin

+++ related-posts-for-wp.php
@@ -104,7 +104,7 @@ class RP4WP {
 		load_plugin_textdomain( 'related-posts-for-wp', false, dirname( plugin_basename( __FILE__ ) ) . '/languages/' );
 
 		// Check if we need to run the installer
-		if ( get_site_option( RP4WP_Constants::OPTION_DO_INSTALL, false ) ) {
+		if ( is_admin() && get_site_option( RP4WP_Constants::OPTION_DO_INSTALL, false ) ) {
 
 			// Delete do install site option
 			delete_site_option( RP4WP_Constants::OPTION_DO_INSTALL );
